generic_tlb: fixing compilation warning in integer expression
 - Using size_t instead of int

git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2574 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/lib/generic_tlb/include/generic_tlb.h
@@ -326,7 +326,7 @@ public:
     {
 
         m_bypass_valid = false;
-        for (int i = 0; i < m_nways * m_nsets; i++) {
+        for (size_t i = 0; i < m_nways * m_nsets; i++) {
             m_valid[i] = false;
 
             // Following inits to avoid potential errors from memory checkers
